Moved VideoStyles into their own file. Updated id for <CandidateCardForListWrapper to be able to be candidate or politician-based.
