+++ src/js/pages/WelcomeForOrganizations.jsx
@@ -5,10 +5,11 @@ import React, { Component, Suspense } from 'react';
 import { Helmet } from 'react-helmet-async';
 import styled from 'styled-components';
 import AnalyticsActions from '../actions/AnalyticsActions';
+import { PlayerContainer, Video } from '../common/components/Style/VideoStyles';
 import historyPush from '../common/utils/historyPush';
 import { renderLog } from '../common/utils/logging';
 import normalizedImagePath from '../common/utils/normalizedImagePath';
-import { BlueTitle, PlayerContainer, SubTitle, Title, Video } from '../components/Welcome/headerWelcomeStyles';
+import { BlueTitle, SubTitle, Title } from '../components/Welcome/headerWelcomeStyles';
 import { Bold, Description, DescriptionContainer, DescriptionImageColumn, DescriptionLeftColumn, Image, NetworkContainer, NetworkImage, Section, SectionTitle, SectionTitleBold } from '../components/Welcome/sectionStyles';
 import Testimonial from '../components/Widgets/Testimonial';
 import AppObservableStore from '../common/stores/AppObservableStore';
